+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 0.0.1
 
 - Initial version.
+
+## 0.0.2
+- Added `construct create` command.
+
+## 0.0.3
+- Added `construct create` command.
+- Added template for mvc.